requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
numpy==2.2.2
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
